v inicial

+++ server/src/main/java/dellemuse/server/importer/BaseImporter.java
@@ -20,82 +20,75 @@ import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
+import dellemuse.model.logging.Logger;
 import dellemuse.server.BaseService;
 import dellemuse.server.Settings;
 import dellemuse.server.db.model.Person;
+import dellemuse.server.db.model.User;
+import dellemuse.server.db.service.DBService;
 import dellemuse.server.db.service.UserDBService;
 import dellemuse.server.objectstorage.ObjectStorageService;
 
 public abstract class BaseImporter extends BaseService implements Importer {
-
-    final String baseDir;
-    
-    final String classDir;
-    
-    @JsonIgnore
-    final File inbox; 
-    
-    @JsonIgnore
-    final File mediaDir; 
-
-    @JsonIgnore
-    final File processed;
+                    
+    static private final Logger logger = Logger.getLogger(BaseImporter.class.getName());
     
     @JsonIgnore
     static private final ObjectMapper importerMapper = new ObjectMapper();
     
-    
-    static  {
+    final String baseDir;
+
+    final String classDir;
+
+    @JsonIgnore
+    final File inbox;
+
+    @JsonIgnore
+    final File mediaDir;
+
+    @JsonIgnore
+    final File processed;
+
+    static {
         importerMapper.registerModule(new JavaTimeModule());
         importerMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
         importerMapper.registerModule(new Jdk8Module());
-        //mapper.configure(SerializationFeature.FAIL_ON_SELF_REFERENCES, false);
+        // mapper.configure(SerializationFeature.FAIL_ON_SELF_REFERENCES, false);
     }
 
     @JsonIgnore
     @Autowired
     private final UserDBService userDBService;
 
-
-    
     @JsonIgnore
     @Autowired
     private final ObjectStorageService objectStorageService;
-    
-    
-    
-    public File getMediaDir() {
-            return this.mediaDir;
-    }
-    
-    
-    public BaseImporter(        Settings settings, 
-                                UserDBService userDBService, 
-                                ObjectStorageService objectStorageService,
-                                String classDir)  {
+
+    public BaseImporter(Settings settings, UserDBService userDBService, ObjectStorageService objectStorageService,
+            String classDir) {
         super(settings);
-        
+
         this.objectStorageService = objectStorageService;
-        this.userDBService=userDBService;
-        
-        this.baseDir=settings.getImporterBaseDir();
-        this.classDir=classDir;
-       
-        this.inbox     = new File ( getBaseDir() + File.separator  + "inbox"        + File.separator + getClassDir());
-        this.processed = new File ( getBaseDir()  + File.separator + "processed"    + File.separator + getClassDir());
-        this.mediaDir   = new File ( getBaseDir() + File.separator  + "inbox"        + File.separator + getClassDir() + "media");
-         
+        this.userDBService = userDBService;
+
+        this.baseDir = settings.getImporterBaseDir();
+        this.classDir = classDir;
+
+        this.inbox = new File(getBaseDir() + File.separator + "inbox" + File.separator + getClassDir());
+        this.processed = new File(getBaseDir() + File.separator + "processed" + File.separator + getClassDir());
+        this.mediaDir = new File(
+                getBaseDir() + File.separator + "inbox" + File.separator + getClassDir() + File.separator + "media");
+
     }
-    
+
     public String getClassDir() {
         return this.classDir;
     }
-    
+
     public String getBaseDir() {
         return this.baseDir;
     }
-    
-    
+
     public File getInbox() {
         return this.inbox;
     }
@@ -104,53 +97,66 @@ public abstract class BaseImporter extends BaseService implements Importer {
         return this.processed;
     }
 
+    public File getMediaDir() {
+        return this.mediaDir;
+    }
 
     @Override
-    @JsonIgnore 
+    @JsonIgnore
     public ObjectMapper getObjectMapper() {
         return importerMapper;
     }
 
-    
     public synchronized void execute() throws IOException {
-        
-        if (!getInbox().exists() || !getInbox().isDirectory()) {
-                FileUtils.forceMkdir(getInbox());
-                return;
-        }
-        
-        
-        for (File file: getInbox().listFiles()) {
+
+        if (!getProcessed().exists() || !getProcessed().isDirectory())
+            FileUtils.forceMkdir(getProcessed());
+
+        if (!getInbox().exists() || !getInbox().isDirectory())
+            FileUtils.forceMkdir(getInbox());
+
+        if (!getMediaDir().exists() || !getMediaDir().isDirectory())
+            FileUtils.forceMkdir(getMediaDir());
+
+        for (File file : getInbox().listFiles()) {
+            
             if (isJson(file)) {
                 boolean success = read(file);
-                if (success)
+                
+                if (success) {
+                    logger.debug( this.classDir + " - " + file.getName() + " -> ok");
                     process(file);
+                }
             }
         }
     }
 
-    
     protected abstract boolean read(File file) throws StreamReadException, DatabindException, IOException;
 
-    
-    protected void process( File file) throws IOException {
-        File destination = new File ( getProcessed() + File.separator + file.getName());
-        Files.move(file.toPath(), destination.toPath(),  StandardCopyOption.REPLACE_EXISTING);
+    protected void process(File file) throws IOException {
+
+        if (file == null || !file.exists())
+            return;
+
+        File destination = new File(getProcessed() + File.separator + file.getName());
+        Files.move(file.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
     }
 
     protected boolean isJson(File file) {
         return file.getName().toLowerCase().endsWith(".json");
     }
 
+    public User findRoot() {
+        return getUserDBService().findRoot();
+    }
+
+    
     public UserDBService getUserDBService() {
         return userDBService;
     }
 
-
     public ObjectStorageService getObjectStorageService() {
         return objectStorageService;
     }
-
-    
 
 }
